Fix route presentation when delegate changes view hierarchy (#5)
This fixes a bug where a route would not be presented because our delegate's implementation of urlRouter:routedUrlWillBePresented:completionHandler: modified the view hierarchy to remove the view controller returned by urlRouter:viewControllerToPresentRoutedUrlFrom:.

This introduces the condition warned of in the interface comment: if urlRouter:viewControllerToPresentRoutedUrlFrom: returns nil, and urlRouter:shouldPresentRoute: returns FSQUrlRouterAllowRouting, the WillBePresented delegate would get called but the route and DidPresent won't get called. I am ok with this being documented undefined behavior.